# services/files/internal/infra/repo/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type File struct {
	ID          pgtype.UUID
	OwnerUserID pgtype.UUID
	Name        string
	Size        int64
	MimeType    string
	Description *string
	UploadedAt  pgtype.Timestamptz
}

type Tag struct {
	ID        pgtype.UUID
	Name      string
	Color     string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}